feat: create database with properties without configuration
Description: This PR includes the basic implements of the Create Notion Database with properties havhing not configuration with expection of Number getting as a default field type number. There is no handling of required parameters but optional params are handled. The Notification once database is created succesfully or not will be configured in addup PR and Type of CreateNotionDatabase under NotionSDK will be created in addup PR too